requirements.txt
fastapi==0.93.0
uvicorn==0.20.0
pydantic==1.10.5
starlette==0.25.0
python-multipart==0.0.6
requests
joblib
torch
pypdf2==3.0.1
tensorflow~=2.11.0
keras~=2.11.0
transformers==4.26.1
huggingface_hub
pytesseract==0.3.10
pdf2image~=1.16.3
pandas~=1.4.4
numpy~=1.23.5
pillow~=9.2.0
h5py~=3.7.0
pathlib~=1.0.1
image-classifiers==1.0.0
httpx
pytest
